chore(front): Move status message component to other directory
