--- client/src/components/layout/type/model.ts
@@ -1,7 +1,0 @@
-import type { ReactNode } from "react";
-
-export type StatusMessgeProps = {
-  code: number;
-  message: string;
-  children: ReactNode;
-};
