+++ client/src/components/status-message/type/model.ts
@@ -1,0 +1,7 @@
+import type { ReactNode } from "react";
+
+export type StatusMessageProps = {
+  code: number;
+  message: string;
+  children: ReactNode;
+};
