+++ client/src/stories/status-message/status-message.stories.tsx
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { ComponentMeta, ComponentStoryObj } from "@storybook/react";
-import StatusMessage from "~/components/layout/statusMessage";
+import StatusMessage from "~/components/status-message";
 
 type T = typeof StatusMessage;
 type Story = ComponentStoryObj<T>;
