+++ client/src/tests/components/status-message/status-message.test.tsx
@@ -1,15 +1,13 @@
 import { composeStories } from "@storybook/testing-react";
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import * as stories from "~/stories/layout/statusMessage.stories";
+import * as stories from "~/stories/status-message/status-message.stories";
 
 const { Default } = composeStories(stories);
 
 describe("(components) header", () => {
   test("test Snap Shot", () => {
-    const { container } = render(
-      <Default code={404} message="page not found" />,
-    );
+    const { container } = render(<Default />);
     expect(container).toMatchSnapshot();
   });
 });
